requirements.txt
Flask
Flask-WTF
discord.py
python-dotenv
gunicorn